Se reorganizó el appsidebar y la reestructura de las carpetas cambiando los nombres usando buenas prácticas
