+++ apps/system/app/[subdomain]/(dashboard)/administracion/ciclos-escolares/SchoolCycleCard.tsx
