+++ apps/system/app/[subdomain]/(dashboard)/administracion/ciclos-escolares/page.tsx
