+++ apps/system/app/[subdomain]/(dashboard)/evaluacion/calificaciones-periodos/page.tsx
@@ -21,9 +21,9 @@ import { TermAverageMatrix } from "./term-average-matrix";
 import { Id } from "@repo/convex/convex/_generated/dataModel";
 import { api } from "@repo/convex/convex/_generated/api";
 import { useQuery, useMutation } from "convex/react";
-import { useUserWithConvex } from "../../../../stores/userStore";
+import { useUserWithConvex } from "../../../../../stores/userStore";
 import { useUser } from "@clerk/nextjs";
-import { useCurrentSchool } from "../../../../stores/userSchoolsStore";
+import { useCurrentSchool } from "../../../../../stores/userSchoolsStore";
 import { toast } from "sonner";
 import { Button } from "@repo/ui/components/shadcn/button";
 import { Input } from "@repo/ui/components/shadcn/input";
