+++ apps/system/app/[subdomain]/(dashboard)/evaluacion/calificaciones-periodos/term-average-matrix.tsx
