+++ apps/system/app/[subdomain]/(dashboard)/general/aulas/page.tsx
