+++ apps/system/app/[subdomain]/(dashboard)/general/grupos/GroupCard.tsx
