+++ apps/system/app/[subdomain]/(dashboard)/general/horarios/scheduleUtils.ts
