+++ apps/system/app/[subdomain]/(dashboard)/general/materias/SubjectCard.tsx
