+++ apps/system/components/menu/search-form.tsx
@@ -1,0 +1,22 @@
+import { Search } from "lucide-react"
+
+import { Label } from "@repo/ui/components/shadcn/label"
+import { SidebarInput } from "@repo/ui/components/shadcn/sidebar"
+
+export function SearchForm({ ...props }: React.ComponentProps<"form">) {
+  return (
+    <form {...props}>
+      <div className="relative">
+        <Label htmlFor="search" className="sr-only">
+          Search
+        </Label>
+        <SidebarInput
+          id="search"
+          placeholder="Type to search..."
+          className="h-8 pl-7"
+        />
+        <Search className="pointer-events-none absolute top-1/2 left-2 size-4 -translate-y-1/2 opacity-50 select-none" />
+      </div>
+    </form>
+  )
+}
